struct: exchange() is a machine-word swap, and the word grew -- it WORKS now
2.dfs.c:147 is a generic swap declared `int *p1,*p2; int temp;'.  Its four
callers pass THREE DIFFERENT THINGS, all four bytes on a VAX:

  2.dfs.c:140   &graph[v], &graph[loo]         VERT ** -- ROW POINTERS
  2.dfs.c:141   &v, &loo                       VERT *  -- vertex numbers
  3.loop.c:136  &graph[temp], &graph[v]        VERT ** -- row pointers
  3.then.c:73   &LCHILD(v,THEN), &..(v,ELSE)   VERT *  -- graph cells

With `int *' it exchanged only the LOW 32 BITS and left both high halves.
negate() swaps THEN=-1 with ELSE=8 and the cells came out

  THEN = 0xFFFFFFFF00000008   low half 8 correct, high half the old -1
  ELSE = 0x00000000FFFFFFFF   low half -1 correct, high half the old 8

DEFINED() is (v >= 0), so the undefined child read as +4294967295 and
mkthen fired.  Declared `long' not VERT: two callers pass POINTERS, so what
upstream means is "one machine word" -- `int' on a VAX, `long' here.  VERT
would be right for two callers and a lie for the other two.  The two
row-pointer sites are worse and were never reached: a half-width swap of two
heap pointers corrupts the graph's row table outright.

IT WAS IN MY OWN SWEEP OUTPUT.  The int* sweep printed
`2.dfs.c:148:int *p1,*p2;' under a heading I wrote as "int* locals", and I
scanned past it looking for locals.  It is a K&R PARAMETER LIST -- the one
context where a declaration alone on a line is not a local.  Two further
rounds of instrumentation re-derived what that line already said.

AND BISECTING BY PHASE FOUND IT IN ONE RUN after two rounds of guessing at
mechanism.  Three fprintfs in 3.main.c:

  BEFORE getreach [6]=-1 [7]=-1
  AFTER  getreach [6]=-1 [7]=-1
  AFTER  getflow  [6]=8  [7]=-1     <-- still CLEAN

which refuted the entire "wrong stride during graph construction" line.  The
row dump before it had corrected me twice too: cell [4] read 4425033912 and
looked like garbage, but it CHANGES BETWEEN RUNS -- ASLR varying it is what
says it is a correctly stored 64-bit pointer, which is what widening VERT
was for.

structure now does its job.  Measured on seven inputs; the four that matter:

  IF..GOTO backwards   -> REPEAT ... UNTIL(!(n .gt. 0))
  DO with GOTO out     -> DO i = 1, n { IF(..) break 1; ... }
  three-way GOTO       -> IF / ELSE IF / ELSE
  computed GOTO, 2 subs-> SWITCH(k) { CASE 1: ... }

Every GOTO gone, replaced by the construct it encoded.  Baseline control
(pre-VERT sources rebuilt in a scratch tree) SIGSEGVs in fixvalue on all
seven, including the trivial one.

Still not in the Makefile in THIS commit -- integration is the next one, and
PORTING.md lists the four steps.  The standing reason not to build it is
discharged.

make test: 2429 passed, 0 failed across 17 suites.

+++ v8/src/cmd/struct/2.dfs.c
@@ -144,10 +144,38 @@ VERT v;
 	FATH(iter) = UNDEFINED;	/* will be defined later along with FATH for DOVX */
 	}
 
+/*
+ * PORT: `long' throughout, was `int'.  This is a GENERIC MACHINE-WORD SWAP and
+ * the machine word grew; it is not a swap of ints that happens to be reused.
+ * Its four callers pass three different things, and on a VAX all three were
+ * four bytes:
+ *
+ *	2.dfs.c:140   &graph[v], &graph[loo]        VERT ** -- ROW POINTERS
+ *	2.dfs.c:141   &v, &loo                      VERT *  -- vertex numbers
+ *	3.loop.c:136  &graph[temp], &graph[v]       VERT ** -- row pointers
+ *	3.then.c:73   &LCHILD(v,THEN), &..(v,ELSE)  VERT *  -- graph cells
+ *
+ * With `int *' it exchanged only the LOW 32 BITS of each object and left both
+ * high halves in place.  Measured on `IF (N .GT. 0) GOTO 10': negate() swaps
+ * THEN=-1 with ELSE=8 and the cells came out
+ *
+ *	THEN = 0xFFFFFFFF00000008   (low half 8 correct, high half the old -1)
+ *	ELSE = 0x00000000FFFFFFFF   (low half -1 correct, high half the old 8)
+ *
+ * so DEFINED() -- which is `v >= 0' -- read the undefined child as the
+ * positive 4294967295 and mkthen's assertion fired.  The two rows that swap
+ * ROW POINTERS are worse and were never reached by these inputs: a half-width
+ * pointer swap corrupts the graph's row table outright.
+ *
+ * `long' rather than VERT because two of the four callers pass pointers, and
+ * what upstream means here is "one machine word" -- which is what `int' was
+ * on a VAX and what `long' is under LP64.  There is no one V8 type spelling
+ * that; VERT would be right for two callers and a lie for the other two.
+ */
 exchange(p1,p2)		/* exchange values of p1,p2 */
-int *p1,*p2;
+long *p1,*p2;
 	{
-	int temp;
+	long temp;
 	temp = *p1;
 	*p1 = *p2;
 	*p2 = temp;
